correções
Algumas correções no que o Wander fez.
