+++ src/img/baralho/copas/6 de copas.jpg
