+++ src/img/baralho/copas/7 de copas.jpg
