+++ src/img/baralho/espadilha/4 de espadilha.jpg
