+++ src/img/baralho/espadilha/6 de espadilha.jpg
