+++ src/img/cenario/placar.jpg
