+++ src/img/vetorizar/moeda-vetorizar.png
